move md code snippets
- update imports
