+++ docs/src/lib/components/GettingStarted.svelte
@@ -1,5 +1,5 @@
 <script>
-  import GettingStartedGuide from './getting-started-guide.md'
+  import GettingStartedGuide from './code-snippet/getting-started-guide.md'
 </script>
 
 <section id="get-started">
