+++ docs/src/lib/components/code-snippet/install-code.md
